requirements-fresh.txt
Django>=4.1,<4.2
aiosmtpd
google-cloud-recaptcha-enterprise
beautifulsoup4